+++ app/controllers/services/information_controller.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+class Services::InformationController < Services::ApplicationController
+  skip_before_action :authenticate_user!
+  before_action :ensure_in_session!
+
+  def show
+    @offer = Offer.find(session[session_key]["offer_id"])
+  end
+end