+++ spec/factories/services.rb
@@ -5,7 +5,7 @@ FactoryBot.define do
     sequence(:title) { |n| "service #{n}" }
     sequence(:description) { |n| "service #{n} description" }
     sequence(:tagline) { |n| "service #{n} tagline" }
-    sequence(:service_type) { :normal }
+    sequence(:service_type) { :orderable }
 
     factory :open_access_service do
       sequence(:connected_url) { "https://sample.url" }
@@ -33,6 +33,10 @@ FactoryBot.define do
 
     after(:create) do |service, _evaluator|
       service.reindex(refresh: true)
+    end
+    factory :external_service do
+      sequence(:service_type) { :external }
+      sequence(:connected_url) { "https://sample.url" }
     end
   end
 end