+++ spec/models/service_spec.rb
@@ -48,8 +48,14 @@ RSpec.describe Service do
     it { is_expected.to validate_presence_of(:connected_url) }
   end
 
-  context "if not open access" do
-    before { allow(subject).to receive(:open_access?) { false } }
+  context "if catalog" do
+    before { allow(subject).to receive(:external?) { true } }
+
+    it { is_expected.to validate_presence_of(:connected_url) }
+  end
+
+  context "if normal" do
+    before { allow(subject).to receive(:orderable?) { true } }
 
     it { is_expected.to_not validate_presence_of(:connected_url) }
   end