+++ CHANGELOG.md
@@ -9,14 +9,26 @@ Please view this file on the master branch, on stable branches it's out of date.
 ## [Unreleased]
 
 ### Added
+- EOSC logo update (@bwilk)
+- Catalog can be added to project (@martaswiatkowska)
+- Project_new_message mails content (@martaswiatkowska)
+- Information step to ordering path (@martaswiatkowska)
 
 ### Changed
+- Internal server error message (@goreck888)
+- Anonymous user can enter service offer selection section (@mkasztelnik)
+- Service access button (@martaswiatkowska)
+- Services can be added to project multiple times if offer has parameters (@martaswiatkowska)
+- Service and offer type enum values (@goreck888)
+- Service ordering summary step unification (@goreck888)
+- Anonymous user can enter information step (@martaswiatkowska)
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+- Active Order button in service preview (@goreck888)
 
 ### Security